Conflicts:
	gen/com/runrunfast/homegym/R.java

+++ src/com/runrunfast/homegym/record/TotalRecord.java
@@ -4,5 +4,5 @@ public class TotalRecord {
 	public float total_kcal;
 	public int total_time;
 	public int total_days;
-	public int total_food;
+	public float total_food;
 }
